+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Fixed
 * SyncManager had some inconsistent locking which could result in data races and/or deadlocks, mostly in ways that would never be hit outside of tests doing very strange things (since v10.0.0).
 * Reduce the peak memory usage of changeset uploading by eliminating an extra copy of each changeset which was held in memory.
+* Don't keep trying to refresh the access token if the client's clock is more than 30 minutes fast. ([#4941](https://github.com/realm/realm-core/issues/4941))
 
 ### Breaking changes
 * None.